feat(blog): Add new blog post about configuring Python version in pyproject.toml
This commit introduces a new blog post that explains 
how to set the minimum and maximum Python versions in `pyproject.toml`. 
The post also discusses the impact of this configuration on `llamabot` 
and the new syntax for indicating the union of types in Python 3.10. 
The new syntax is visually easier to understand, which can make coding more efficient.