feat(blog): update and enhance 'coding on an ipad' blog post
This commit introduces several updates to the 
'coding on an ipad with codespaces and blink' blog post. 
It adds more tags related to the content of the post, 
such as 'jupyter notebook', 'keybindings', 'macbook air', 
'nano', 'productivity', 'syntax highlighting', 'terminal', 'tmux', and 'vscode'. 
It also reorders some of the existing tags. 
Additionally, it expands the summary section with more detailed information 
about the content and the author's experience.