feat(notebooks): Add new llamabot notebook
This commit introduces a new Jupyter notebook called 'llamabot.ipynb'. This notebook contains code for a blogging bot that can compose LinkedIn and Twitter posts. The bot uses BeautifulSoup to scrape text from a webpage and then uses this text to compose social media posts. The notebook also includes code for loading the bot, sending requests to a webpage, and parsing the response.