feat(blogbot_scraper): Add initial version of blogbot scraper
This commit introduces a new Jupyter notebook, blogbot_scraper.ipynb, which contains code to scrape blog URLs from a given website. The code uses BeautifulSoup to parse the HTML and extract the relevant links. The links are then composed into a list of strings and a dictionary mapping links to titles is created. 

This is the initial version of the scraper and further improvements and optimizations may be needed in the future.