feat(blog): Add new blog post on service vs product-oriented data science
This commit introduces a new blog post titled "Service vs. Product-Oriented Data Science". The post explores the two flavors of data science work in biotech: service-oriented and product-oriented. It discusses their value, challenges, and the choice between them depending on the situation. The author advocates for a product-first orientation, using service-first to demonstrate the value of the tools we build.