feat(blogbot): add functionality to format tags
Add functionality to format tags for a blog post in the BlogBot API. The tags are generated based on the input text and are returned as a JSON object with the key "post_text" and the value as the tags concatenated as one string with line breaks. This allows for easier parsing and usage of the tags. The tags are limited to a maximum of 10 and should have a maximum of two words, all lowercase, and no "#" symbol. Spaces are allowed in the tags. This change improves the usability and flexibility of the BlogBot API.

[END COMMIT MESSAGE]