feat: Add blog post on automatically writing git commit messages
This commit adds a new blog post titled "How to automatically write git commit messages".
The blog post discusses the challenges of writing informative commit messages and introduces the Conventional Commits specification as a helpful guideline.
It also explains how the LlamaBot Pythonic interface to Large Language Models (LLMs) can be used to automatically generate commit messages based on the git diff.
The post provides instructions on installing the prepare-commit-msg hook to run the LlamaBot after pre-commit hooks and before editing the commit message.
The author emphasizes the importance of precision and clarity when interacting with LLMs.