feat(blogbot): Add support for Substack posts
This commit adds support for composing Substack posts in the blog bot.
It includes a new function `compose_substack_post` in the `prompts.py` file.
This function prompts the user to compose a post for Substack that entices readers
to read the blog post and includes a call to action.
The function returns the composed post text in JSON format.

The commit also includes some minor code formatting changes in the `api.py` file.