feat: Add blog post on choosing a (conda) distribution of Python
This commit adds a new blog post titled "How to choose a (conda) distribution of Python".
The blog post discusses the differences between the Anaconda, Miniconda, and Miniforge distributions of Python.
It provides information on the features, package managers, and default packages included in each distribution.
The blog post also offers recommendations on which distribution to use based on different needs and preferences.