docs(blog): Add tip on consistent installation directory for conda distributions
This commit adds a tip to the blog post on choosing a conda distribution of Python.
The tip explains how all distributions can be installed into the `~/anaconda` directory for consistent behaviour,
regardless of the installation source.
It also provides an example of using the `-p` flag to specify the installation directory.