feat(blog): Add new blog post about creating a local pre-commit hook for image resizing
This commit introduces a new blog post titled "How I made a local pre-commit hook to resize images".
The post details the process of creating a pre-commit hook that automatically resizes images within a repository, specifically targeting logos.
The hook is written in Python and uses the pre-commit framework.
The post also includes a discussion on the potential of distributing these hooks for wider use.
A corresponding logo image for the blog post is also added.