feat(blog): add new blog post and update image resizing script (#92)
* feat(blog): add new blog post and update image resizing script

This commit introduces a new blog post titled "Elevating Team Performance: Feedback Strategies for Data Science Leaders". The post discusses various strategies for providing effective feedback in a data science team. It also includes a new logo in .webp format for the blog post.

In addition, the image resizing script has been updated to resize images to a maximum width of 1024px, up from the previous 800px. The blog post template has also been updated to support .webp format logos.

This commit aims to provide valuable insights for data science leaders and improve the visual quality of blog post logos.

* docs(blog): enhance explanation of 'in situ' in feedback strategies post

In the blog post about feedback strategies for data science leaders, the term 'in situ' is now better explained.
A link to the Wikipedia page for 'in situ' has been added for readers who want to understand the term in more depth.
The changes aim to improve the clarity of the post and provide more context for readers unfamiliar with the term.

* feat(blog): add support for webp format in blog logo

This commit introduces the ability to use the webp format for blog logos, in addition to the existing png format.
If a webp logo is available, it will be used; otherwise, the png logo will be used.

* fix(blog): ensure image always displays in blog posts

This commit moves the image tag outside the conditional statement to ensure that the image always displays in blog posts,
regardless of whether the image is a .png or .webp file.

* fix(blog.html): correct image rendering condition

This commit fixes the placement of the endif statement in the image rendering condition.
This ensures that the image is only rendered when the image source is available, preventing potential errors or broken image links.