feat(blog): Add new blog post about GitHub Actions secrets
This commit introduces a new blog post titled "GitHub Actions secrets need to be explicitly declared". The post discusses the author's experience with setting up environment variables in GitHub Actions, specifically the need to explicitly declare secrets in the workflow. The post also includes a new logo image for the blog post.