feat(blog): Add new blog post about setting up Ollama server on home network
This commit introduces a new blog post titled "LlamaBot with Ollama on my home virtual private network". The post details the process of setting up an Ollama server on a home private network, using a Tailscale virtual private network and a GPU box running Ubuntu Linux. It also explains how to install Ollama on the GPU box, configure it for network access, and test its remote access. The post further demonstrates how to interact with the Ollama server using LlamaBot's SimpleBot class and highlights the limitation that Ollama needs to have downloaded a model before it can be used from SimpleBot. The blog post concludes by stating the benefits of using Ollama and LlamaBot in extending the usable life of an idle GPU box.

In addition to the blog post, a new logo image file has also been added to the blog post's directory.